backend/requirements.txt
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.1
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.5
django-environ==0.4.5
django-extensions==2.2.9
django-redis==4.11.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==1.4.0
googlemaps==4.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
locationiq @ git+https://github.com/location-iq/locationiq-python-client.git@88b9b1dbf4bd47814cb59af558907a913fa2ccca
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
numpy==1.18.4
pandas==1.0.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
ratelimit==2.2.1
raven==6.9.0
redis==3.4.1
requests==2.23.0
rope==0.16.0
simplegeneric==0.8.1
six==1.14.0
sqlparse==0.3.1
testfixtures==6.14.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
whitenoise==5.0.1
